requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-storage==2.10.0
google-genai>=0.3.0
google-cloud-aiplatform>=1.38.0  # Required for vertexai (used by existing gemini_service.py)
python-dotenv==1.0.0

# Core runtime (DB / templates)
sqlalchemy==2.0.23
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5

# Media processing with enhanced video support
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=10.1.0
numpy>=1.24.0,<2.0
imageio>=2.31.1
imageio-ffmpeg>=0.4.8

# Optional heavy features (removed for now): mediapipe, scikit-learn, pandas, nltk, pytesseract,
# python-docx, reportlab, fpdf2, imageio, imageio-ffmpeg, moviepy, aiofiles, asyncio-extras
# If needed later, re-add from requirements.backup.txt.

# Utilities
requests==2.31.0
urllib3>=1.25.4,<2.1
typing-extensions==4.8.0
psutil==5.9.6  # System metrics for Phase 9 testing

# NOTE: Removed google-cloud-aiplatform (Vertex AI) to avoid resolver explosion.
#       Gemini は google-genai (新公式 SDK) を利用。旧 google-generativeai は非推奨につき削除。

# Phase 4: RAG System & Enterprise Features
boto3==1.34.0  # AWS S3 for reference materials and PDFs
elasticsearch==8.11.0  # Semantic search
redis==5.0.1  # Message broker for Celery
celery==5.3.4  # Async task processing
flower==2.0.1  # Celery monitoring

# Document Processing
PyPDF2==3.0.1  # PDF text extraction
pdfplumber==0.10.3  # Enhanced PDF parsing
python-docx==1.1.0  # Word document processing
openpyxl==3.1.2  # Excel file processing

# PDF Generation (Phase 6)
weasyprint==60.1  # HTML to PDF conversion

# Markdown Processing (Phase 6 - Unified Manual Generator)
markdown==3.5.1  # Markdown to HTML conversion

# Testing
faker==20.1.0  # Test data generation

gunicorn==22.0.0
waitress==3.0.0
pytest==7.4.3
